pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vibevoice"
version = "0.1.0"
description = "Voice-to-text using local Whisper model with keyboard input"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Marc Päpper" }
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.23.2",
    "faster-whisper==1.1.1",
    "ctranslate2==4.5.0",
    "torch==2.4.0+cu124",
    "torchaudio==2.4.0+cu124",
    "python-multipart==0.0.7",
    "python-dotenv==1.0.0",
    "sounddevice==0.4.6",
    "numpy>=1.26.0",
    "requests==2.32.3",
    "pynput==1.7.6",
    "scipy==1.11.3"
]

[project.scripts]
vibevoice = "vibevoice.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vibevoice"]
